Check pointers are non-NULL before checking their values
When a pointer is returned by a function check that it is
non-NULL before checking its value.

This is a resolution to the problem reported in Debian Bug #716429
See: https://bugs.debian.org/716429

That bug report exercises the problem of the return value
of the call to vanessa_socket_server_bindv() in
vanessa_socket_server_connectv().

Signed-off-by: Simon Horman <horms@verge.net.au>